Add prefix/directory support in etcd_get.
We parse this into a list of newline-separated keys.  It looks a little funny
in etcd-test, and it's a bit inefficient because we don't include the values
even though we have them right in our hands, but maybe it's still better than
getting nothing at all.  Also, since the YAJL tree interface doesn't seem to
have any reasonable way to iterate over the elements of a JSON array, we have
to "cheat" by looking inside the yajl_val structure ourselves.  This is likely
to break some day, but maybe by then the necessary API will have been added.

Signed-off-by: Jeff Darcy <jdarcy@redhat.com>